created friendcard component renamed details card to game card
