+++ src/components/emotion/styledCard.js
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const StyledCard = styled.div`
+  background-color: #f5f5f5;
+  padding: 0px 12px;
+  border: none;
+  border-radius: 8px;
+  box-shadow: 5px 5px 7px #d9d9d9, -5px -5px 7px #ffff;
+  margin: 10px;
+`;
+
+export default StyledCard;
